# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.taskResponse:
    properties:
      avatarId:
        type: integer
      avatarName:
        type: string
      createdAt:
        type: string
      errorMessage:
        type: string
      id:
        type: integer
      outputVideoS3Url:
        type: string
      scriptText:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: LingCast 数字人直播平台 API（api-admin / api-user / api-scheduler 共用一份 OpenAPI
    文档；管理端写操作与 /users、/admin/* 需登录会话，观众端与 Worker Webhook 公开）。
  title: LingCast API
  version: 1.0.0
paths:
...
  /tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List broadcast tasks (newest first)
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: avatarId + scriptText
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.taskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create a broadcast (offline video) task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete a broadcast task
      tags:
      - tasks
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.taskResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Poll task status and output URL
      tags:
      - tasks
  /tasks/{id}/retry:
    post:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.taskResponse'
      summary: Re-enqueue a failed task
      tags:
      - tasks
  /tasks/{id}/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: status + outputVideoS3Url + error
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 'Worker webhook: task status callback'
      tags:
      - worker
  /tts/preview:
    post:
      consumes:
      - application/json
      parameters:
      - description: voiceId + text (<=200 chars)
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - audio/mpeg
      responses:
        "200":
          description: audio/mpeg bytes
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Voice audition (MP3 bytes, proxied to tts-service)
      tags:
      - tts
...
swagger: "2.0"